command/format: pass declared cmd context subset to Format object
This subset is declared with 'context_specs' and is being built on the
metaclass level in the top-down, extending (additive/non-subtractive)
manner (just like the set of methods is growing monotonously top-down
the hierarchy).

Currently, only 'validator_specs' (mapping protocols to the somewhat
abstract "validator spec", '' key being used as a failback) is of
interest this way, and can be used as an on-off validation toggle
on user's request (by convence, '' value stands for off state).

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>