Reconsider plugin (class/symbol) to module organization
This is a preparation work for being able to pick up plugin of choice
nicely without the necessity of scanning the whole name/path space
followed with some initialization routines across all plugins even if
they won't be used for sure.  Such heavy-weight approach is fine for
using clufter as CLI tool (covering potentially any available use-case),
not so much for the use as a library with presumably a priori determined
task to perform.

As generally one module can (and often is practical!) contain several
modules, we are introducing following convention to allow our proposed
filtering to have a good false-positives/convenience ratio:

    module shall contain only plugins the name of which contains
    the same prefix as the module's filename does, where the prefix
    is assumed the part before first occurrence of '_'
    ('-' for stringified/decorated plugin names)

As a concrete example, formats/ccs.py module contains ccs and ccs_flat
plugins, hence when ccs_flat is to be picked with this fine granularity
and using default search path, only formats/ccs.py file will match
the prefix test and hence only ccs and ccs_flat will be prepared
for further use (one false positive in this case is a result of the
mentioned trade-off).

This commit normalizes the plugin-module allocation to follow
the convention, reflects these changes in tests, etc. and adds a note
to the HACKING doc file.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>