filters/ccsflat2pcs: encode services into valid, yet uncomplete CIB
IOW, there will have to be another pass that will use the data we have
prepared by known, massage them, and finally (and hopefully) spit out
the desired full-fledged cib, incl. reasonable constraints arrangement.

Apparently, ./run-check doesn't work as of this commit, but it's
worthless to edit it just to keep temporarily passing, as the final
result is different anyway.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>