filter (XSLT): make nested comments survive repeated "XSLT"
This is imporant is case of "<descent-mix> - <descent>" structural
pattern, in which the original xsl:comment gets dissolved into
<!-- comment --> with the first XSLT run, which then disappear
completely upon subsequent runs.

Now it is preserved up the last point possible -- gets translated
upon postprocess that is (only) applied on the final product.
Btw. we utilize the fact in DOM (at least as seen with lxml),
text content of element node is handled the same in case of comment
node, so there no more magic needed.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>