facts: add ability to translate logical cmd to distro cmd
The target distro cmd is expected to keep the assumed template form
(say, a command to install packages should contain "{commands}" field
that is then substituted, via "format" method, with the desired
packages, completing the command to be used further in the process).

Unlike with logical packages, nothing is considered a (suitable)
distro-wide default.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>