ccs-flatten: avoid proper RA *.metadata confusion with own ones
Previously, running the CLI without proper installation on the machine
with some /usr/share/cluster/*.metadata files present (typically RHEL 6
with resource-agents installed) meant that "onvenient fallback for
local/test deployment" hadn't been activated, leading to various errors
in the conversion due to most of the *.metadata files missing anyway
(most notably for "script").

Now we take an advantage of own/embedded metadata files ending with
".sh.metadata", therefore we are not confused in this way (at least
not so easily, but good enough for the time being).

Reported-by: Chris Feist <cfeist@redhat.com>
Signed-off-by: Jan Pokorný <jpokorny@redhat.com>