filters/ccsflat2pcsprelude[mysql]: fix additional_parameters
When listen_address parameter of the rgmanager's version of mysql RA
is contributing to additional_parameters parameter of hearbeats's one,
we don't want to accidentally inject unexpected mysql invocation params
as a side effect -- that's why the quoting was in place since beginning.
But quotes have to be around whole "word" in shell's view so the
unquoting will take place at the right occasion during agent invocation.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1219572
Reported-by: Patrik Hagara <phagara@redhat.com>
Signed-off-by: Jan Pokorný <jpokorny@redhat.com>