filter: fix issue possibly workarounded with text-only descent
IIRC, this is related to "filters/cib2pcscmd[group]" case where usage of
a parameter (presumably propagated from def_first) was attempted but did
not really worked with original descent-mix (hence text-only descent was
deemed a quick fix).  This patch should be a much better solution to
that very problem ... so flip these "clufter:descent" instructions as
well.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>