commands/ccs2pcscmd-flatiron: fix a "guidance"/shell thinko
Original code would not work for more occurrences of common pairs --
only the status of last evaluation would then be considered for general
failure case.  Instead, short-circuit in case of OK case and end the
loop with failure (false) otherwise -- which will then be propagated
as a general loop status upon the end of non-short-circuited loop.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>