utils: new "hybridmethod" decorator
Similar to hybridproperty: first parameter of the decorated method
is flexibly bound to either the instance or the class depending
on the context of invokation.  Effectively, the only difference
is that hybridmethod is expected to be passed arguments, unlike
with property where it is called directly (w/o parameters) upon
access.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>