run-check: --dist goes prior alias cmd to influence it
This hasn't been considered before (hence the bug being fixed), and
it seems best if --dist in "global options" (prior to cmd spec) is the
only to influence command alias resolution, whereas --dist as "command
option" can influence some specifics for the command itself.
This way, natural causality on the command line is preserved.

Alternatively, we could look-ahead if there is dist defined as command
option and considered it in the alias resolution.  But rather than
that, it would more bearable to remove --dist/--sys from command options
altogether but that would prevent injecting them when command-by-symlink
is exercised, so do not do this either.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>