filters/cmd_wrap: revisit group/line wrap-filling algo
Previous one was just splitting into lines within a single group
(also, this was limited to at most one split! that in turn made
it a pretty broken text wrapping).  Now it is two-ways allowing
also short-enough groups to be put on the same line, and as many
line splits is made as necessary for a proper text wrapping.

Also update the test counterparts (function run_check_cmd of
run-check script succeeds again after quite a long time!).

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>