filters/cib2pcscmd[resources]: terminate if "template" observed
+ point the user to the root cause.  It makes no sense and could be
pretty misleading when we cannot produce templates via pcs while these
are vital for the dependent resources to work properly.

True to be told, template + resource(s) can always be expressed as
fully-fledged resource (with parameters transferred down to resource
following the same [intuitive] logic) but that would defeat why this
provision was introduced in the first place, wouldn't it?

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>