commands: memoize/refactor filter in chains
Note that previosly, there was an unintended behavior of filter
chain definition containing a single named block like in:

    (ccsflat2cibfinal_chain,
        ('cib-revitalize',
            ('cib-meld-templates',
                ('cib2pcscmd',
                    ('stringiter-combine2'  # , ('cmd-wrap' ...
                     )))))

worked, despite this accidentally led to the expected behavior,
i.e., the same as if the explicit tail sequence (starting with
'cib-revitalize') was appended directly as a sibling of the last
item within ccsflat2cibfinal_chain.  Note that abusing this
"feature" is highly discouraged, it's really a bug and, moreover,
it will stop working when you try to apply such nesting recursively
more than once.

Just of a curiosity of when something behaves a bit better than
ever thought of (while making things proportionally worse later on).

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>