filters/ccsflat2cibprelude[rm]: forget about undeployed resources
This was actually a conversion bug in that these unused (just declared)
resources were promoted into ones effectively executed under new stack.

Alternative solution would be to keep them included as it used to be
plus set "is-managed" meta-attribute for them to "false", but there is
at least one deeper underlying issue: ccs_flatten is not interested in
these unemployed items in the "resources" stanza at all (they are not
validated, nor they are paired with metadata-provided defaults, etc.).

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>