filters/cib2pcscmd[rsc_colocation]: saner score handling
...wrt. pcs treatment of missing score within the command line
arguments (=INFINITY) vs. pacemaker treatment of missing score
in the CIB (=0, no influence).

Futhermore, the original idea of emitting concise output where
possible (such as when score is [+]INFINITY) is preserved.

As a consequence, id specification of the constraint needs to be
left at the end, but that is actually welcome as consistent with
how rsc_ticket constraint is handled in clufter.

Also, add respective unit tests.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>