filters/needleqdevicexml2pcscmd[device*]: emit only when supported
Otherwise, if there is quorum device configured, emit a warning
it cannot be configured with particular older pcs+corosync pertaining
the target distribution.  There is also (still aliased/ambiguous)
warning when its net model is configured and that's not supported
with respective pcs+corosync combo.

Due to the new arrangement, unit test for needleqdevicexml2pcscmd need
to be passed the target distribution explicitly.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>