formats/command: get_merged: fix nested commands vs. shlex.split
Apparently, shlex.split is supposed to be parsing just simple, flat
commands, not a richer shell language provided syntax, especially wrt.
command/process substitution (not handling included quoting properly)
and (command) list separators (of which only ';' is being handled
as of this commit to prevent it being considered a part of a word
and subsequently enquoted as a whole), so we just add one more custom
parsing layer atop for these.

Also drop the misleading comment that reflects an initial intention
but not how the code eventually evolved.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>