commands/*2pcscmd: emit suitable shell in a shebang
Beside allowing the outputs to be marked executable (and consequently
run with a correct interpreter), but also to suggest the user which
shell it should be run with.  This is especially important as
{ccs,pcs}2pcscmd commands without --noguidance (or --dry-run) switches
would generate shell code containing process substitutions, which is
incompatible with generic POSIX shell (also /bin/sh) and the reported
syntax errors may confuse users.

Resolves: <https://bugzilla.redhat.com/1381531>
Reported-by: Miroslav Lisik <mlisik@redhat.com>
Signed-off-by: Jan Pokorný <jpokorny@redhat.com>