completion: better dealing with '=' value to option assignment
...in generated bash completion.  Previously, trying to set any option
with a standard "-[-]option=value" syntax while pressing [TAB] after
having typed '=' would barf with

> -bash: declare: cannot use `-f' to make functions

because '=' would be unguarded part of the function name being checked
for existence.  Solution is to skip such treatment early as with any
alleged option (starting with '-').  Furthermore, '=' is now
automatically suggested as an optional ending of options that are
supposed to accept a value argument.

Resolves: <https://bugzilla.redhat.com/1381565>
Reported-by: Miroslav Lisik <mlisik@redhat.com>
Signed-off-by: Jan Pokorný <jpokorny@redhat.com>