command: terminal filters can peek at resolved IO declarations
This is to eventually get rid of unnecessarily premature decisions
leading to needless duplication of suboptimal code, which is run too
early (filters were so far unaware of the real output declaration [which
is now being changed a bit] so it had to be done at command level)
without wider knowledge to make any better reasoning.

The showcase currently is detecting whether the command will hit the
terminal (as in isatty(3)) with some of its output streams, which was
oversimplified as checking whether this particular output was specified
as '-' (designating stdout per convention + actual stdout is a terminal)
during the command mapping it's input to the io chain to be processed
phase.

At the downside, that resolution now happens en masse prior to
watterfalling through the chain, no longer ad-hoc prior to execution
current filter, which eliminates a possibility of high dynamism where
one filter directly influences the output IO declaration of another
based on its own execution, but this hasn't been useful thus far
and hardly will.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>