facts: infer_comp: avoid undesired number-to-string comparison
This boils down to a misfeature of Python 2 fixed in Python 3 (porting
to which discovered this otherwise masked issue).  The root cause is
that the per-distro "branch" sequences mix several types of data and
where original code assumed just a tuple-encoded version information,
it could be a plain string custom-encoding the "extras".  Now we guard
against misusing the later where the former is expected.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>