2to3: pep-3115: PY2 metaclass syntax ignored in PY3
New syntax was designated in PEP 3115 -- Metaclasses in Python 3000:
https://www.python.org/dev/peps/pep-3115/
(It would be possible to leave the existing __metaclass__ syntax in
place. [that's not the case, though])

The new syntax is too different from the old one, so any conceivable way
to combine both comes unwieldy.  Fortunately, we can just emulate the
affect of creating new class through the means of meta class in a
similar fashion a decorator can be emulated by explicit decoration
function pass-through.  This solution is courtesy of
https://gitorious.org/flufl-enum/flufl-enum/source/master?p=flufl-enum:flufl-enum.git;a=blob;f=flufl/enum/_enum.py;hb=4.0.1#l184
that was referred to in one of the Python porting documents.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>