3to3: pep-0328: non-dot imports considered absolute in PY3
This in itself was not established through PEP, but is documented:
https://docs.python.org/3.0/whatsnew/3.0.html#removed-syntax
(The only acceptable syntax for relative imports is from .[module]
import name. All import forms not starting with . are interpreted
as absolute imports. (PEP 0328).)

Note that ./run-tests context may be fine (due to being run with
modified sys.path), but there's a real issue in PY3 if the test in
question was run as a main module (it's a desired aspect of the test
suite's design as it is handy to run particular tests like this at
times).

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>