2to3: execfile is gone in PY3
Reference:
https://docs.python.org/3.0/whatsnew/3.0.html#builtins

Note that "exec(open(fn).read())" as suggested is not a full-fledged
substitute, as the source path is not handed over for introspection
purposes (incl. traceback messages) this way, but it's trivial to make
sure this piece of information is absorbed as well through another
building -- compile -- that is actually passed also the file name.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>