2to3 follow-up: drop pep-3149 extra ABI versioning for an executable
Back to basics, just use the logic as in the concurrent
distutils code without the extension adding part rather
than delegate-to-ancestor-and-postprocess.  The resulting
binary is an ABI-insensitive executable (we are just having
a hackery in place to reuse module compilation recipe so
as to produce this other type of binary) anyway.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>