command: ensure deterministic assignment of short options
...in respect to the long ones.  Now the listing order of respective
long options in per-command doctext is followed, except for the long
ones starting with the same initial letter -- such groups are, in full,
appended to the list under construction in place where the first
occurrence of its member would be added per the stated principle.
This is also the only case pre-/post-commit results could differ
at the end of the day, i.e., in the final assignment of short
options to their long counterparts.

Note this actually doubles as an improvement over the previous state
where the processing order was basically undefined (subject of the
order in which given python implementation + possibly the executed
instance iterates over the dictionary in _figure_parser_desc_opts).

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>