command: prevent secondary clashes on short options
"Secondary clashes" refer to subsequent round of attempts to find
an unoccupied short option (using the other letters of its long
form in some priority order) when the first letter is already
claimed.  Previously the logic was to weak to prevent two options
conflicting on the same letter picked like this (coincidentally,
it hadn't happened yet, but I've discovered the issue when adding
new, experimental option).

Also add an "extraneous parameter passing" comment for possible
consideration in the future.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>