format: make swallow method pay a bigger attention to validations
Previously, it would only care if the representation of the format
instance to be stored was validatable.  This neglience was circumventing
the idea of "what's been constructed successfully is assuredly valid".

It becomes even more pressing because we want to introduce the concept
of "one-way upgradability of the format" (treating multiple similar
versions in the same family of formats uniformly, with implicit
on-demand upgrades where required), and it's not reasonably doable
in case of CIB without keeping tabs on the validity properly
(as the lack thereof is an impulse to try upgrading, similarly
to what pacemaker does under the hood).

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>