format: split (while sanitizing) etree_rng_validator
Previously, "XML.validator_specs[ETREE]" (as of the closure at
the time of parsing the method at hand, i.e., that value as specified
precisely at XML class, whereas there was an expectation of polymorphism
ever since -- resolving to the context-of-use value instead) would
be used as a default specification for the RelaxNG-based validator,
statically boiling down to using arbitrary *.rng file at the path scope
("root" variable) of particular descendant of XML class.

This is the main issue addressed with factoring the respective part
of XML.etree_rng_validator into XML._etree_rng_validator_specs
internal-only method.  The other justification is for the upcoming
helper to figure out just the subset of the files matching the
possibly (and usually) glob-pattern prescribed specifications
that, at the same time, don't match the specifications of any
of the class descendants.

(Also, flip the branching within XML._etree_rng_validator_specs for
better clarity.)

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>