format: Format: resort to __new__ calling out __init__ equivalent
This is because of the upcoming inter-format arrangement to allow for
(one-way) format upgrades.  As a first step, we add __new__ method
that currently only performs more-or-less the same as the implicit
one (with an outlook this will be made to possibly, upon direct
validation failure, try to instantiate using each of the upgradable
descendants in row, then to upgrade it to the destined one).
Unfortunately, we need to have a firm grip on running __init__
equivalent from within this __new__ method, so we rename current
__init__ methods to _construct that we can have a full control
over, as opposed to __init__ being run independently by the runtime
machiner as long as __new__ returns an instance exactly of the
expected class (which may not always be the case, based on __init__
only prototype).

Note that __init__ is still here should it be handy for anything
on top of what's delivered with _construct.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>