format: XML: add etree_rng_validator_proper_specs method
Useful for working with validation specs (presumably glob-patterns)
that pertains just the particular class, without any overlap with
the class descendants.  This is useful for upgrades with formats
that somehow encodes the format version itself -- for instance, it
may be reasonable if the field gets set to the minimal version of
class proper.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>