filters/cib2pcscmd[acls]: restrict per where ACLs supported
...meaning a target system where two conditions are met: (a) pacemaker
supports CIB schema 2.0 (at least), and (b) pcs itself already supports
the "pcs acl ..." syntax.

The slavishly repeated guards at particular acl elements (target, group,
role) within the schema-reflecting tree discovered a slight design
weakness with <clufter:descent-mix> approach to hierarchical composition
of the transformation handling because it would be preferred if these
very could be wrapped in a guard at once, but this would require the
tree descending algoritm being taught some new branching construct.
That might be pretty convoluted, hence just add a note to KNOWN-ISSUES
file.

Also, drop the pacemaker-1.1.12 and "acls" association, as that is
misleading -- pacemaker could enforce some kind of ACL enforcement
since version 1.1.5 -- because what was really meant to be expressed
was "supports new style of ACL encoding within CIB", but it is
(silently) implied by the support of that 2.0 version of schema.
This change imposes a slight modification of one of the unit tests.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>