clufter.spec: factor "borrow schemas from pacemaker" out to a script
It wasn't economical to carry all the current-at-pacemaker-release CIB
validation schemas, as there are clear update/versioning rules[1]:
- non-conflicting addtions: keep the same major version
- conflicting changes: bump the major version, craft the update path
  (as XSLT)
and hence everything in X.Y is subsumed with X.Y+1 incl. some proper
additions.  Therefore we carry just the latest minor version for each
major one, and because of locally established (not necessarily true
for pacemaker) update-just-to-lowest-minor-per-new-major rule
(usually X.Y -> X+1.0), we add also these lowest minor counterparts,
to validate the sanity of the upgrade if not for anything else.
As an exception, we avoid pacemaker-1.0 schema as the lowest common
denominator we have in mind is pacemaker-1.2 and we don't support
upgrade from pre-1.0 schema version, anyway.

[1] https://github.com/ClusterLabs/pacemaker/tree/Pacemaker-1.1.17-rc3/xml#updating-schema-files

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>