tests/_com: make CommonFilterTestCase base class determinism-friendly
Some filters will intentionally spray the outputs with a limited amount
of randomness, which is desirable in standard use, but not when the
outputs are to be compared in the test scenarios.  Currently what
decides whether to avoid this pseudorandom non-determism in the
generated outputs is the "NOSALT" instrumentation as a respective
environment variable.

Hence make CommonFilterTestCase base test case class shared across the
test suite (directly or by inheritance) inject such variable and revert
this back in setUp and tearDown methods, respectively.

To add insult to injury, there's another obstacle because defining that
variable just when the start in question is about to begin may be too
late as the "walk" utilized in the filter may have been precached
from previous tests where the variable hasn't been defined -- particular
modules being visited are imported with artificial module addressing,
and unlike any other module discovery (plugins), repeated visits of
these modules won't, intentionally (avoiding futile execution in some
cases), lead to module reloads.  We could surely apply catch-all
solution but that's not a systemic approach should we ever want to
distinguish on test-by-test basis (as we are exercising here with
CommonFilterTestCase from get go), hence we rather roll out a new
"test case" subclass, DeterministicFilterTestCase, instrumented to
drop this "walk" cache in its setUp method.  Plus some more tweaks
in plugin_registry.py were needed to support this new arrangement.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>