filters/needleqdevicexml2pcscmd[device*]: add support for heuristics
There's a slight possibility that pcs syntax is not yet set in stone,
but let's neglect that for now.  Also speculatively, assume the
prerequisites to apply that -- support in both corosync and pcs -- is
present in RHEL 7.5, which also enables the new test case to be
exercised at all.

Also refresh corosync.rng with the new configuration items to set
heuristics up, which is a prerequisite for pcs2pcscmd-needle command
to handle that out of the box.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>