facts: aliases_rel/cluster_map/versions_extra: support Fedora Rawhide
For our purpose, it's not optimal to treat Rawhide as the nearest
upcoming (to be) branched version, because so far we are strictly
adhering to the rule of "track lowest versions of package present in
particular distro version, even if they may be updated later on", and
for that, it would be messy to combine two different approaches within
single-label distro version -- newest version till the branching,
oldest version from that point on.

Hence, as Rawhide is currently progressing towards release 28, we are
simply rolling out new "27.999" (beside getting obvious "rawhide" alias
also artificially aliased as "28" because that's how Rawhide identifies
itself at that point and we want to support this gracefully) entry that
will keep tracking newest versions as of modification time, till 28 is
branched, then it will get relabelled as "28.999" (together with bumping
the alias to 29) while 28 will get a proper entry on its own to be
updated till this release is out so as to truthfully reflect the
versions at that very release-ready point (to be carried as oldest
ones on board since then on).

Also reflect that both pacemaker 1.1.18 and Fedora 27 have gone live
already.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>