facts+run_check: treat "system" part as indeed a dynamic factor
Let's make supported_dists a function of the "sys" input.
Handy, should we want to introduce support for non-linux systems.

In the same vein, also make run-check full roundtrip tests operate
with a fully qualifiede target, i.e., make sure to specify the system
part that was omitted so far.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>