main: parser_callback_dist: make it translate release aliases, too
Previously, just a distro name was a subject of normalization.  Note
that it's vital if the filters are ever aboutto work with those values
directly because ambiguity would be expensive (not that it'd be
a very good idea to avoid reusable abstractions, but still).
We are only intersted in normalizing the first, most significant part
of the version information, usually carrying the actual numeric version,
while the rest encodes code names and whatnot.

Also note the we are still limited with what the originating
"main: extra --dist handling callback" commit mentioned, i.e.,
the run-time-dynamic defaults are still carried on without
this normalization in place.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>