main: translate distro + release from run-time-dynamic defaults as well
Tackling the debt mentioned in the previous commit, we factor out the
translation logic into _resolve_dist function that is now called also
for said detected defaults.   It also means that, for instance,
autodetected Fedora Rawhide gets properly normalized to 27.999 version
as of this commit, as filters using these raw values (not the best idea
as likewise mentioned in the preceding commit) should only rely on
normalized data if they really need to at all.

Note that distribution names are assurably lower case already, but now
also the detected default gets alias-resolved for sure.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>