facts+main: add preliminary support for FreeBSD
It appears that, not so long ago, FreeBSD has jumped on the bandwagon
and started to ship corosync and pacemaker in its "ports" collection:
- corosync: https://svnweb.freebsd.org/ports?view=revision&revision=411764
- pacemaker: https://svnweb.freebsd.org/ports?view=revision&revision=411799

Hence it's not a bad idea to track their versions as updated across
distro releases.  To make that to be useful out-of-the-box when run on
that very platform, also tweak the system/distro preconfiguration of the
defaults, relying on "/bin/freebsd-version -u" result to identify the
particular distro version, as there doesn't appear to be any
standardized text files indicating the same reliably.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>