formats/corosync.rng: support nodelist.node.name configuration option
This option within corosync configuration was silently established (and
suggested to overcome certain node identification issues) by pacemaker
without any awareness within the former, and hence was completely
foreign also to a detached attempt to formalize this public config,
interface (known as "corosync.conf" file) by the means of RelaxNG
schema describing its XML equivalent and materialized as an in-tree
corosync.rng file.  This used to be prone to validation issues with the
cib2pcscmd* commands, but since corosync now acknowledges that option
just as any other native one[*], set the schema to follow suit.

[*] https://github.com/corosync/corosync/pull/286

Resolves: <https://bugzilla.redhat.com/1517834>
Signed-off-by: Jan Pokorný <jpokorny@redhat.com>