clufter.spec: make remaining setuptools case python2-qualified for Py2
This is a follow-up for "clufter.spec: make setuptools and lxml" commit.
For distros not catching up with Python version qualification of the
packages around Python 2 -> 3 shift, this change may need to be skipped,
but that's not assumed likely.

Related-change: <https://src.fedoraproject.org/rpms/clufter/c/e101a0b>
Signed-off-by: Jan Pokorný <jpokorny@redhat.com>